requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
Pillow
open3d
transformers
torch
timm
opencv-python
coremltools; sys_platform == 'darwin'